backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
fastapi==0.121.3
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
ftfy==6.3.1
grad-cam==1.5.5
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.6
numpy==2.2.6
open_clip_torch==3.2.0
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
PyMuPDF==1.26.5
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.22
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.1
ttach==0.0.3
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.3.230
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
lightgbm>=4.0
shap>=0.44
lime>=0.2
pandas>=2.0
imbalanced-learn>=0.11
PyJWT==2.10.1
psycopg[binary]==3.2.12
SQLAlchemy==2.0.44
groq>=0.9
python-dotenv>=1.0
pydicom==3.0.2
pytest==8.4.2
